+++ examples/readme_table_no_tty.rs
@@ -1,5 +1,4 @@
-use comfy_table::presets::UTF8_FULL;
-use comfy_table::*;
+use comfy_table::{presets::UTF8_FULL, *};
 
 // This example works even with the `tty` feature disabled
 // You can try it out with `cargo run --example no_tty --no-default-features`