+++ src/style/mod.rs
@@ -5,7 +5,8 @@ mod cell;
 mod color;
 mod column;
 /// Contains modifiers, that can be used to alter certain parts of a preset.\
-/// For instance, the [UTF8_ROUND_CORNERS](modifiers::UTF8_ROUND_CORNERS) replaces all corners with round UTF8 box corners.
+/// For instance, the [UTF8_ROUND_CORNERS](modifiers::UTF8_ROUND_CORNERS) replaces all corners with
+/// round UTF8 box corners.
 pub mod modifiers;
 /// This module provides styling presets for tables.\
 /// Every preset has an example preview.
@@ -23,8 +24,7 @@ pub use table::{ContentArrangement, TableComponent};
 /// Convenience module to have cleaner and "identical" conditional re-exports for style enums.
 #[cfg(all(feature = "tty", not(feature = "reexport_crossterm")))]
 mod styling_enums {
-    pub use super::attribute::*;
-    pub use super::color::*;
+    pub use super::{attribute::*, color::*};
 }
 
 /// Re-export the crossterm type directly instead of using the internal mirrored types.
@@ -32,9 +32,11 @@ mod styling_enums {
 /// project with different versions, but may also be very convenient for developers.
 #[cfg(all(feature = "tty", feature = "reexport_crossterm"))]
 mod styling_enums {
-    /// Attributes used for styling cell content. Reexport of crossterm's [Attributes](crossterm::style::Attribute) enum.
+    /// Attributes used for styling cell content. Reexport of crossterm's
+    /// [Attributes](crossterm::style::Attribute) enum.
     pub use crossterm::style::Attribute;
-    /// Colors used for styling cell content. Reexport of crossterm's [Color](crossterm::style::Color) enum.
+    /// Colors used for styling cell content. Reexport of crossterm's
+    /// [Color](crossterm::style::Color) enum.
     pub use crossterm::style::Color;
 
     /// Convenience function to have the same mapping code for reexported types.