+++ tests/all/padding_test.rs
@@ -1,10 +1,9 @@
+use comfy_table::*;
 use pretty_assertions::assert_eq;
 
-use comfy_table::*;
-
-#[test]
 /// Columns can set a custom padding.
 /// Ensure these settings are working.
+#[test]
 fn custom_padding() {
     let mut table = Table::new();
     table